requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
google-generativeai==0.3.2
gunicorn==21.2.0
scrapy==2.11.0
pandas==2.1.4
numpy==1.26.2